break up actions.ts
